requirements.txt
Django>=5.2.3
djangorestframework>=3.14.0
drf-spectacular>=0.27.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.5
coverage>=7.0.0
ruff>=0.1.0
